requirements.txt
# Finalized requirements.txt

# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0

# Image Processing
opencv-python==4.8.0.76
Pillow==9.5.0              # RESOLVED CONFLICT: Changed from 10.0.0 for streamlit==1.26.0
scikit-image==0.21.0
spectral==0.23.1

# Web Framework
streamlit==1.26.0          # CHOICE: Selected Streamlit

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
pyyaml==6.0.1
tqdm==4.66.1
python-dotenv==1.0.0

# Report Generation
dicttoxml==1.7.16
lxml==4.9.3

# Optional (Uncomment if needed)
# h5py==3.9.0